+++ NEWS.md
@@ -5,6 +5,7 @@
 
 ## Major Changes
 * Added the gene_update parameter to runTBsigProfiler() to allow users to check signature/sample gene names for excel mogrified or outdated gene symbols
+* Added Chendi_HIV_2 signature (8/24)
 
 ## Minor Changes
 * Rewrote runTBsigProfiler() to be shorter and easier to maintain